Redirects should be installed together with pages
