+++ wcc-contentful-app/lib/generators/wcc/templates/page/models/redirect.rb
